Set the environment again
