+++ hello-react/src/App.css
@@ -1,2 +1,5 @@
 @import '~antd/dist/antd.css';
 
+body{
+  background: black;
+}
