+++ hello-react/src/components/main/Slider.js
@@ -1,0 +1,11 @@
+import React, { Component } from "react";
+
+class Center extends Component {
+  render() {
+    return (
+      <div></div>
+    );
+  }
+}
+
+export default Center;
