--- hello-react/src/style.css
@@ -1,5 +1,0 @@
-body{
-  background: black;
-  color:white;
-  font-size: 50px;
-}
